Fix png compressed text and itxt handling
iTXt isn't in the same format as zTXt -- the body is in three parts (language, translated keyword, and maybe-compressed text) rather than one. FIx up the code to read it properly.

zTXt and iTXt may be larger than the size of the temporary buffer. If so then allocate a temporary temporary buffer, since crashing with a buffer-too-big error is kind of embarrassing.

Add in a newpng  test image that has some text associated with it.

Decode (hopefully) the comment blocks in a gif image, and dump them out in our metadata dumper program. Note that I don't have a file that actually has any comments in it so I don't know if this works or not.