Plumb contexts through png and implement timeout checks
Plumb the context through the PNG package, or at least through most of it.

Add in a check at the beginning of the generic chunk processor to see if the context has expired. Return an error if so. This is the initial part of the timeout checks, though they should be added to more parts of the code. (Especially to the decompression code, which can potentially take degenerate time in the face of malicious input)

Add in a test to make sure that basic cancellation failure happens.