Add a few ICC base types
The ICC standard clearly defines a bunch of base types that it uses which don't directly correspond to Go's standard types. Define them so we can use them later.

This is necessary to avoid potential conversion loss between Go types and ICC types, even if it's a bit annoying to deal with. (For example the ICC defines several fixed-point decimal types)