Expunge Decode
Switch a bunch of the code in the image packages from using legacy format-specific Decode to using the new DecodeExtended.

Add in some convenience structs since the current inline struct way of specifying options is a bit awkward.

Add in DecodeImage and DecodeMetadata to the base image package that decode the image or metadata. These are akin to the Decode and DecodeConfig calls, except they take contexts and ReadOption lists.