Kill Config
Turns out that with proper metadata support there's no reason that Config even really needs to be a thing -- whats in there is just a small subset of metadata. So, nuke all references to it except a few necessary for backwards compatibility, and add guaranteed height, width, and color model fields, along with a metadata-guaranteed GetConfig accessor for convenience, to stuff.