Write the iCCP, tEXt, and zTXt chunks.
We can't set the ICCP chunk but we can at least round-trip it.

We can now write the tEXt and zTXt chunks, and there are tests to see if we can read the tEXt (but not the zTXt) chunk.

Fix a bug in tEXt reading that the writer code exposed.

(#7)