Fix ztxt reader, and write ztxt out
Compression type 0 is flate, not uncompressed, and there is no compression type 1. Or any other type, really. So fix that.

The writer code was also broken. That was awesome.

Added a test which uncovered both issues. Remember, kids, features don't really exist until there's a test to prove it!

(Addressing #7)