ICCP chunks can be > 768 bytes
Allocate a new temporary buffer if we need to read more than 768 bytes for an ICCP chunk. Which is pretty likely, honestly, since ICC profies are pretty big.

Add in some test data that has an ICC profile > 768 bytes. (Specifically a pic I took on vacation -- Kauai'i is lovely, tbh)

#7 applies here