Read and write XMP data
XMP entries are just special magic-named iTXt chunks, so look for the magic name and pull the data out if we find it.

Teach the writing code how to write out XMP data if it's attached. This will handle either undecoded read-from-disk data or data that's been attached to the metadata, though at the moment there's no way to actually attach XMP data since that module's a stub.

touches #7, #10, and #11