Some metadata data definitions
Add a bunch of exif data field and note their tag numbers. (These should probably be pointers to things but we'll leave that for the moment)

Make the xmp things pointers to the structs so we can mark them as optional, though reading the XMP standard again and looking at what gets dumped out in various programs I suspect this will all need to be redone. Also XML is a horrible format for pretty much anything and you shouldn't use it unless forced.