Add in APPx metadata extraction, sort of
Add code to extract and save the various APPx chunks from a jpeg file. We can't decode most of them, but we can at least hold on to them for the moment.

Explicitly extract out the ICC color profile. That we can recognize and can pass to the generic ICC decoder. (Which doesn't work, but thats a separate issue)

Add in a short list of tags we're likely to want to decode. (These are the tags that OS X Photos puts in its jpeg files)