Decode the jpeg thumbnail, if there is one
Whether there's a thumbnail or not depends on the x and y size -- if they're 0 then there's no thumbnail. Which is annoying, but ancient formats gonna do what they do for space reasons. At least it doesn't indicate no-thumbnail by having a short buffer.

Decode the thumbnail if it exists. It feels like there ought to be a more efficient way to do this, but for the moment loops will do.